Created project folder
Made project folder. Created "surface" terrain. Raised one part of the ground and made a hole for the cave entrance. Trying to apply a two-sided terrain to the shader, so that the material is visible on the "inside" of the cave instead of being transparent. I got a two-sided shader to work on a Plane in an earlier assignment, but I'm struggling trying to apply it to the Terrain.