First Build
Tested movement system on Oculus headset. Need to refine starting position. Grabbing board works but is unwieldy. Need to adjust so that board is grabbed at end rather than at center. Will also be more manageable with a smaller gap and board. Also going to add a teleport system to the maze floor blocks to allow that as a control scheme.