Trying to figure out a way to have grabbing and teleporting
Having a hell of a time separating the grab and teleport functions.
Two acceptable outcomes:
-Make one BUTTON grab and another BUTTON teleport.
-Make one HAND grab and the other HAND teleport.
While spending a few hours messing with this, I inadvertently deleted the right hand. I brought in a new XR rig to replace the old one and re-added all the settings I'd put on the other one. It's working again, but I still haven't figured out differentiating grab and teleport controls. Found a tutorial that may help:

https://youtu.be/mh1lYjSS9es