Added an Ax object
I added an ax to the scene which you can test your script on.

You should also take a look at the changes we made to the door raise script. As it was before, it wasn't making a perceptible change when you stepped on the mat. Maybe you needed to add a loop? Krystof switched it to make an abrupt movement rather than a gradual one so we could test it, so it has the abrupt change right now.
Other thoughts:
When I made a moving platform, the XR Rig didn't "stick to it" (stay on top of it while it moved). Maybe you should just make walls that crush you (respawn you when they make contact). You could make crushers that come upward, downward, and from the side.