Testing new axe trap in 12-8 adding elements back in
I put an instance of your new trap by the entry of the maze.
I am working in the scene "Maze Test Scene 12-8 adding elements back in." This is the scene that has the Maze Master animations at the beginning and end.
I think I copied over everything from the scene you were working in (11-29). The pressure plate causes the axe to drop, but right now it is not injuring me.
Please take a look and see if there's something I didn't copy over right.
Something to consider, since time is running out: What if we just make the axe kill you, the same way the lava does?